Refactor common functionality to base class
