+++ src/main/php/net/daringfireball/markdown/Code.class.php
@@ -1,16 +1,6 @@
 <?php namespace net\daringfireball\markdown;
 
-class Code extends Node {
-  public $value;
-
-  /**
-   * Creates a new code element
-   *
-   * @param string $value
-   */
-  public function __construct($value) {
-    $this->value= $value;
-  }
+class Code extends ValueNode {
 
   /**
    * Emit this node
@@ -21,25 +11,5 @@ class Code extends Node {
    */
   public function emit($emitter, $definitions= []) {
     return $emitter->emitCode($this, $definitions);
-  }
-
-  /** @return string */
-  public function toString() {
-    return nameof($this).'<'.$this->value.'>';
-  }
-
-  /** @return string */
-  public function hashCode() {
-    return '`'.md5($this->value);
-  }
-
-  /**
-   * Returns whether a given comparison value is equal to this node list
-   *
-   * @param  var $value
-   * @return string
-   */
-  public function compareTo($value) {
-    return $value instanceof self ? strcmp($this->value, $value->value) : 1;
   }
 }
